refactor: Extract data loading logic into class (#913)
